backend/requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
python-dotenv==1.0.0
google-generativeai>=0.3.1
aiohttp>=3.8.0,<3.9.0
pydantic==1.10.8
requests==2.31.0